% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwPackage.R
\docType{package}
\name{hwPackage}
\alias{hwPackage}
\alias{hwPackage-package}
\title{hwPackage package (this is the title)}
\description{
longer description in text of what the package is being 
assebled to address (description)
}

