% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iomReview.R
\name{iomReview}
\alias{iomReview}
\title{our basic data frame iomReview}
\format{this is a 43 obs. of 14 variables data frame.}
\source{
I downloaded it from Module3 SISBID repository on GitHub.
}
\description{
this really does not do anything
}

